refact: Service 인터페이스와 구현체 분리
